requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
scikit-learn
faiss-cpu
requests